requirements-test.txt
click
pytest
pyyaml
requests
